+++ execution_engine/src/core/engine_state/engine_config.rs
@@ -8,6 +8,8 @@ pub const DEFAULT_MAX_QUERY_DEPTH: u64 = 5;
 pub const DEFAULT_MAX_ASSOCIATED_KEYS: u32 = 100;
 /// Default value for maximum runtime call stack height configuration option.
 pub const DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT: u32 = 12;
+/// Default value for maximum StoredValue serialized size configuration option.
+pub const DEFAULT_MAX_STORED_VALUE_SIZE: u32 = 8 * 1024 * 1024;
 
 /// The runtime configuration of the execution engine
 #[derive(Debug, Copy, Clone)]
@@ -19,6 +21,7 @@ pub struct EngineConfig {
     /// [`Weight`](casper_types::account::Weight)s) for a single account.
     max_associated_keys: u32,
     max_runtime_call_stack_height: u32,
+    max_stored_value_size: u32,
     wasm_config: WasmConfig,
     system_config: SystemConfig,
 }
@@ -29,6 +32,7 @@ impl Default for EngineConfig {
             max_query_depth: DEFAULT_MAX_QUERY_DEPTH,
             max_associated_keys: DEFAULT_MAX_ASSOCIATED_KEYS,
             max_runtime_call_stack_height: DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
+            max_stored_value_size: DEFAULT_MAX_STORED_VALUE_SIZE,
             wasm_config: WasmConfig::default(),
             system_config: SystemConfig::default(),
         }
@@ -41,6 +45,7 @@ impl EngineConfig {
         max_query_depth: u64,
         max_associated_keys: u32,
         max_runtime_call_stack_height: u32,
+        max_stored_value_size: u32,
         wasm_config: WasmConfig,
         system_config: SystemConfig,
     ) -> EngineConfig {
@@ -48,6 +53,7 @@ impl EngineConfig {
             max_query_depth,
             max_associated_keys,
             max_runtime_call_stack_height,
+            max_stored_value_size,
             wasm_config,
             system_config,
         }
@@ -61,6 +67,11 @@ impl EngineConfig {
     /// Returns the current max runtime call stack height config.
     pub fn max_runtime_call_stack_height(&self) -> u32 {
         self.max_runtime_call_stack_height
+    }
+
+    /// Returns the current max runtime call stack height config.
+    pub fn max_stored_value_size(&self) -> u32 {
+        self.max_stored_value_size
     }
 
     /// Returns the current wasm config.