+++ execution_engine/src/core/engine_state/genesis.rs
@@ -781,7 +781,10 @@ where
         let key = Key::Account(system_account_addr);
         let value = { StoredValue::Account(virtual_system_account.clone()) };
 
-        tracking_copy.borrow_mut().write(key, value);
+        let _ =
+            tracking_copy
+                .borrow_mut()
+                .write(key, value, executor.config().max_stored_value_size());
 
         GenesisInstaller {
             genesis_config_hash,
@@ -801,28 +804,29 @@ where
     }
 
     pub(crate) fn create_mint(&mut self) -> Result<ContractHash, GenesisError> {
-        let round_seigniorage_rate_uref =
-            {
-                let round_seigniorage_rate_uref = self
-                    .address_generator
-                    .borrow_mut()
-                    .new_uref(AccessRights::READ_ADD_WRITE);
+        let round_seigniorage_rate_uref = {
+            let round_seigniorage_rate_uref = self
+                .address_generator
+                .borrow_mut()
+                .new_uref(AccessRights::READ_ADD_WRITE);
 
-                let (round_seigniorage_rate_numer, round_seigniorage_rate_denom) =
-                    self.exec_config.round_seigniorage_rate().into();
-                let round_seigniorage_rate: Ratio<U512> = Ratio::new(
-                    round_seigniorage_rate_numer.into(),
-                    round_seigniorage_rate_denom.into(),
-                );
+            let (round_seigniorage_rate_numer, round_seigniorage_rate_denom) =
+                self.exec_config.round_seigniorage_rate().into();
+            let round_seigniorage_rate: Ratio<U512> = Ratio::new(
+                round_seigniorage_rate_numer.into(),
+                round_seigniorage_rate_denom.into(),
+            );
 
+            let _ =
                 self.tracking_copy.borrow_mut().write(
                     round_seigniorage_rate_uref.into(),
                     StoredValue::CLValue(CLValue::from_t(round_seigniorage_rate).map_err(
                         |_| GenesisError::CLValue(ARG_ROUND_SEIGNIORAGE_RATE.to_string()),
                     )?),
+                    self.executor.config().max_stored_value_size(),
                 );
-                round_seigniorage_rate_uref
-            };
+            round_seigniorage_rate_uref
+        };
 
         let total_supply_uref = {
             let total_supply_uref = self
@@ -830,12 +834,13 @@ where
                 .borrow_mut()
                 .new_uref(AccessRights::READ_ADD_WRITE);
 
-            self.tracking_copy.borrow_mut().write(
+            let _ = self.tracking_copy.borrow_mut().write(
                 total_supply_uref.into(),
                 StoredValue::CLValue(
                     CLValue::from_t(U512::zero())
                         .map_err(|_| GenesisError::CLValue(TOTAL_SUPPLY_KEY.to_string()))?,
                 ),
+                self.executor.config().max_stored_value_size(),
             );
             total_supply_uref
         };
@@ -868,9 +873,10 @@ where
             partial_registry.insert(HANDLE_PAYMENT.to_string(), DEFAULT_ADDRESS.into());
             let cl_registry = CLValue::from_t(partial_registry)
                 .map_err(|error| GenesisError::CLValue(error.to_string()))?;
-            self.tracking_copy.borrow_mut().write(
+            let _ = self.tracking_copy.borrow_mut().write(
                 Key::SystemContractRegistry,
                 StoredValue::CLValue(cl_registry),
+                self.executor.config().max_stored_value_size(),
             );
         }
 
@@ -1028,12 +1034,13 @@ where
             .address_generator
             .borrow_mut()
             .new_uref(AccessRights::READ_ADD_WRITE);
-        self.tracking_copy.borrow_mut().write(
+        let _ = self.tracking_copy.borrow_mut().write(
             era_id_uref.into(),
             StoredValue::CLValue(
                 CLValue::from_t(INITIAL_ERA_ID)
                     .map_err(|_| GenesisError::CLValue(ERA_ID_KEY.to_string()))?,
             ),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(ERA_ID_KEY.into(), era_id_uref.into());
 
@@ -1041,12 +1048,13 @@ where
             .address_generator
             .borrow_mut()
             .new_uref(AccessRights::READ_ADD_WRITE);
-        self.tracking_copy.borrow_mut().write(
+        let _ = self.tracking_copy.borrow_mut().write(
             era_end_timestamp_millis_uref.into(),
             StoredValue::CLValue(
                 CLValue::from_t(INITIAL_ERA_END_TIMESTAMP_MILLIS)
                     .map_err(|_| GenesisError::CLValue(ERA_END_TIMESTAMP_MILLIS_KEY.to_string()))?,
             ),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(
             ERA_END_TIMESTAMP_MILLIS_KEY.into(),
@@ -1057,11 +1065,12 @@ where
             .address_generator
             .borrow_mut()
             .new_uref(AccessRights::READ_ADD_WRITE);
-        self.tracking_copy.borrow_mut().write(
+        let _ = self.tracking_copy.borrow_mut().write(
             initial_seigniorage_recipients_uref.into(),
             StoredValue::CLValue(CLValue::from_t(initial_seigniorage_recipients).map_err(
                 |_| GenesisError::CLValue(SEIGNIORAGE_RECIPIENTS_SNAPSHOT_KEY.to_string()),
             )?),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(
             SEIGNIORAGE_RECIPIENTS_SNAPSHOT_KEY.into(),
@@ -1070,10 +1079,11 @@ where
 
         for (validator_public_key, bid) in validators.into_iter() {
             let validator_account_hash = AccountHash::from(&validator_public_key);
-            self.tracking_copy.borrow_mut().write(
+            let _ = self.tracking_copy.borrow_mut().write(
                 Key::Bid(validator_account_hash),
                 StoredValue::Bid(Box::new(bid)),
-            )
+                self.executor.config().max_stored_value_size(),
+            );
         }
 
         let validator_slots = self.exec_config.validator_slots();
@@ -1081,12 +1091,13 @@ where
             .address_generator
             .borrow_mut()
             .new_uref(AccessRights::READ_ADD_WRITE);
-        self.tracking_copy.borrow_mut().write(
+        let _ = self.tracking_copy.borrow_mut().write(
             validator_slots_uref.into(),
             StoredValue::CLValue(
                 CLValue::from_t(validator_slots)
                     .map_err(|_| GenesisError::CLValue(VALIDATOR_SLOTS_KEY.to_string()))?,
             ),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(VALIDATOR_SLOTS_KEY.into(), validator_slots_uref.into());
 
@@ -1094,12 +1105,13 @@ where
             .address_generator
             .borrow_mut()
             .new_uref(AccessRights::READ_ADD_WRITE);
-        self.tracking_copy.borrow_mut().write(
+        let _ = self.tracking_copy.borrow_mut().write(
             auction_delay_uref.into(),
             StoredValue::CLValue(
                 CLValue::from_t(auction_delay)
                     .map_err(|_| GenesisError::CLValue(AUCTION_DELAY_KEY.to_string()))?,
             ),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(AUCTION_DELAY_KEY.into(), auction_delay_uref.into());
 
@@ -1107,12 +1119,13 @@ where
             .address_generator
             .borrow_mut()
             .new_uref(AccessRights::READ_ADD_WRITE);
-        self.tracking_copy.borrow_mut().write(
+        let _ = self.tracking_copy.borrow_mut().write(
             locked_funds_period_uref.into(),
             StoredValue::CLValue(
                 CLValue::from_t(locked_funds_period_millis)
                     .map_err(|_| GenesisError::CLValue(LOCKED_FUNDS_PERIOD_KEY.to_string()))?,
             ),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(
             LOCKED_FUNDS_PERIOD_KEY.into(),
@@ -1124,12 +1137,13 @@ where
             .address_generator
             .borrow_mut()
             .new_uref(AccessRights::READ_ADD_WRITE);
-        self.tracking_copy.borrow_mut().write(
+        let _ = self.tracking_copy.borrow_mut().write(
             unbonding_delay_uref.into(),
             StoredValue::CLValue(
                 CLValue::from_t(unbonding_delay)
                     .map_err(|_| GenesisError::CLValue(UNBONDING_DELAY_KEY.to_string()))?,
             ),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(UNBONDING_DELAY_KEY.into(), unbonding_delay_uref.into());
 
@@ -1187,7 +1201,11 @@ where
                 main_purse,
             ));
 
-            self.tracking_copy.borrow_mut().write(key, stored_value);
+            let _ = self.tracking_copy.borrow_mut().write(
+                key,
+                stored_value,
+                self.executor.config().max_stored_value_size(),
+            );
         }
 
         Ok(())
@@ -1325,16 +1343,20 @@ where
             contract_package
         };
 
-        self.tracking_copy.borrow_mut().write(
+        let _ = self.tracking_copy.borrow_mut().write(
             contract_wasm_hash.into(),
             StoredValue::ContractWasm(contract_wasm),
+            self.executor.config().max_stored_value_size(),
         );
-        self.tracking_copy
-            .borrow_mut()
-            .write(contract_hash.into(), StoredValue::Contract(contract));
-        self.tracking_copy.borrow_mut().write(
+        let _ = self.tracking_copy.borrow_mut().write(
+            contract_hash.into(),
+            StoredValue::Contract(contract),
+            self.executor.config().max_stored_value_size(),
+        );
+        let _ = self.tracking_copy.borrow_mut().write(
             contract_package_hash.into(),
             StoredValue::ContractPackage(contract_package),
+            self.executor.config().max_stored_value_size(),
         );
 
         (contract_package_hash, contract_hash)
@@ -1361,9 +1383,10 @@ where
         partial_registry.insert(contract_name.to_string(), contract_hash);
         let cl_registry = CLValue::from_t(partial_registry)
             .map_err(|error| GenesisError::CLValue(error.to_string()))?;
-        self.tracking_copy.borrow_mut().write(
+        let _ = self.tracking_copy.borrow_mut().write(
             Key::SystemContractRegistry,
             StoredValue::CLValue(cl_registry),
+            self.executor.config().max_stored_value_size(),
         );
         Ok(())
     }