+++ execution_engine/src/core/runtime_context/mod.rs
@@ -7,16 +7,16 @@ use std::{
     rc::Rc,
 };
 
-use tracing::error;
+use tracing::{error, warn};
 
 use casper_types::{
     account::{
         Account, AccountHash, ActionType, AddKeyFailure, RemoveKeyFailure, SetThresholdFailure,
         UpdateKeyFailure, Weight,
     },
-    bytesrepr::ToBytes,
+    bytesrepr::{ToBytes, U8_SERIALIZED_LENGTH},
     contracts::NamedKeys,
-    system::auction::EraInfo,
+    system::auction::{EraInfo, SeigniorageRecipientsSnapshot},
     AccessRights, BlockTime, CLType, CLValue, Contract, ContractHash, ContractPackage,
     ContractPackageHash, DeployHash, DeployInfo, EntryPointAccess, EntryPointType, Gas, Key,
     KeyTag, Phase, ProtocolVersion, PublicKey, RuntimeArgs, StoredValue, Transfer, TransferAddr,
@@ -28,7 +28,7 @@ use crate::{
         engine_state::{execution_effect::ExecutionEffect, EngineConfig, SystemContractRegistry},
         execution::{AddressGenerator, Error},
         runtime_context::dictionary::DictionaryValue,
-        tracking_copy::{AddResult, TrackingCopy, TrackingCopyExt},
+        tracking_copy::{AddResult, TrackingCopy, TrackingCopyExt, WriteResult},
         Address,
     },
     shared::{execution_journal::ExecutionJournal, newtypes::CorrelationId},
@@ -514,9 +514,12 @@ where
     /// Write a transfer instance to the global state.
     pub fn write_transfer(&mut self, key: Key, value: Transfer) {
         if let Key::Transfer(_) = key {
-            self.tracking_copy
-                .borrow_mut()
-                .write(key, StoredValue::Transfer(value));
+            // Writing a `Transfer` will not exceed write size limit.
+            let _ = self.tracking_copy.borrow_mut().write(
+                key,
+                StoredValue::Transfer(value),
+                self.engine_config.max_stored_value_size(),
+            );
         } else {
             panic!("Do not use this function for writing non-transfer keys")
         }
@@ -525,9 +528,12 @@ where
     /// Write an era info instance to the global state.
     pub fn write_era_info(&mut self, key: Key, value: EraInfo) {
         if let Key::EraInfo(_) = key {
-            self.tracking_copy
-                .borrow_mut()
-                .write(key, StoredValue::EraInfo(value));
+            // Writing an `EraInfo` for 100 validators will not exceed write size limit.
+            let _ = self.tracking_copy.borrow_mut().write(
+                key,
+                StoredValue::EraInfo(value),
+                self.engine_config.max_stored_value_size(),
+            );
         } else {
             panic!("Do not use this function for writing non-era-info keys")
         }
@@ -850,10 +856,14 @@ where
         let bytes_count = stored_value.serialized_length();
         self.charge_gas_storage(bytes_count)?;
 
-        self.tracking_copy
-            .borrow_mut()
-            .write(key.into(), stored_value);
-        Ok(())
+        match self.tracking_copy.borrow_mut().write(
+            key.into(),
+            stored_value,
+            self.engine_config.max_stored_value_size(),
+        ) {
+            WriteResult::Success => Ok(()),
+            WriteResult::ValueTooLarge => Err(Error::ValueTooLarge),
+        }
     }
 
     /// Writes data to a global state and charges for bytes stored.
@@ -867,7 +877,36 @@ where
         self.validate_writeable(&key)?;
         self.validate_key(&key)?;
         self.validate_value(&stored_value)?;
-        self.metered_write_gs_unsafe(key, stored_value)?;
+        self.metered_write_gs_unsafe(key, stored_value)
+    }
+
+    /// Writes a `SeigniorageRecipientsSnapshot` to global state and charges for bytes stored.
+    ///
+    /// The value is force-written, i.e. accidentally exceeding the write size limit will not cause
+    /// this to return `Err`.
+    pub(crate) fn metered_write_gs_seigniorage_recipients_snapshot(
+        &mut self,
+        key: Key,
+        snapshot: SeigniorageRecipientsSnapshot,
+    ) -> Result<(), Error> {
+        let cl_value = CLValue::from_t(snapshot).map_err(Error::from)?;
+        let stored_value = StoredValue::CLValue(cl_value);
+        self.validate_writeable(&key)?;
+        self.validate_key(&key)?;
+        self.validate_value(&stored_value)?;
+
+        let bytes_count = stored_value.serialized_length();
+        self.charge_gas_storage(bytes_count)?;
+
+        let computed_trie_leaf = U8_SERIALIZED_LENGTH
+            .saturating_add(key.serialized_length())
+            .saturating_add(stored_value.serialized_length());
+        if computed_trie_leaf > self.engine_config.max_stored_value_size() as usize {
+            warn!(%key, serialized_length=%stored_value.serialized_length(), "wrote a seigniorage recipients snapshot which is too large");
+        }
+        self.tracking_copy
+            .borrow_mut()
+            .force_write(key, stored_value);
         Ok(())
     }
 
@@ -882,16 +921,18 @@ where
         let value_bytes_count = value.serialized_length();
         self.charge_gas_storage(value_bytes_count)?;
 
-        match self
-            .tracking_copy
-            .borrow_mut()
-            .add(self.correlation_id, key, value)
-        {
+        match self.tracking_copy.borrow_mut().add(
+            self.correlation_id,
+            key,
+            value,
+            self.engine_config.max_stored_value_size(),
+        ) {
             Err(storage_error) => Err(storage_error.into()),
             Ok(AddResult::Success) => Ok(()),
             Ok(AddResult::KeyNotFound(key)) => Err(Error::KeyNotFound(key)),
             Ok(AddResult::TypeMismatch(type_mismatch)) => Err(Error::TypeMismatch(type_mismatch)),
             Ok(AddResult::Serialization(error)) => Err(Error::BytesRepr(error)),
+            Ok(AddResult::ValueTooLarge) => Err(Error::ValueTooLarge),
         }
     }