+++ execution_engine_testing/tests/src/test/regression/mod.rs
@@ -35,6 +35,7 @@ mod gh_1470;
 mod gh_1688;
 mod gh_1902;
 mod gh_2280;
+mod gh_2605;
 mod gov_116;
 mod gov_74;
 mod gov_89_regression;