+++ execution_engine_testing/tests/src/test/system_costs.rs
@@ -4,8 +4,8 @@ use once_cell::sync::Lazy;
 use casper_engine_test_support::{
     utils, DeployItemBuilder, ExecuteRequestBuilder, InMemoryWasmTestBuilder,
     UpgradeRequestBuilder, DEFAULT_ACCOUNTS, DEFAULT_ACCOUNT_ADDR, DEFAULT_ACCOUNT_INITIAL_BALANCE,
-    DEFAULT_ACCOUNT_PUBLIC_KEY, DEFAULT_MAX_ASSOCIATED_KEYS, DEFAULT_PAYMENT,
-    DEFAULT_PROTOCOL_VERSION, DEFAULT_RUN_GENESIS_REQUEST,
+    DEFAULT_ACCOUNT_PUBLIC_KEY, DEFAULT_MAX_ASSOCIATED_KEYS, DEFAULT_MAX_STORED_VALUE_SIZE,
+    DEFAULT_PAYMENT, DEFAULT_PROTOCOL_VERSION, DEFAULT_RUN_GENESIS_REQUEST,
 };
 use casper_execution_engine::{
     core::engine_state::{
@@ -194,6 +194,7 @@ fn upgraded_add_bid_and_withdraw_bid_have_expected_costs() {
         DEFAULT_MAX_QUERY_DEPTH,
         new_max_associated_keys,
         DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
+        DEFAULT_MAX_STORED_VALUE_SIZE,
         WasmConfig::default(),
         new_system_config,
     );
@@ -433,6 +434,7 @@ fn upgraded_delegate_and_undelegate_have_expected_costs() {
         DEFAULT_MAX_QUERY_DEPTH,
         new_max_associated_keys,
         DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
+        DEFAULT_MAX_STORED_VALUE_SIZE,
         WasmConfig::default(),
         new_system_config,
     );
@@ -873,6 +875,7 @@ fn should_verify_wasm_add_bid_wasm_cost_is_not_recursive() {
         DEFAULT_MAX_QUERY_DEPTH,
         new_max_associated_keys,
         DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
+        DEFAULT_MAX_STORED_VALUE_SIZE,
         new_wasm_config,
         new_system_config,
     );