+++ types/src/system/auction/detail.rs
@@ -31,8 +31,7 @@ where
 {
     let key = provider.named_keys_get(name).ok_or(Error::MissingKey)?;
     let uref = key.into_uref().ok_or(Error::InvalidKeyVariant)?;
-    provider.write(uref, value)?;
-    Ok(())
+    provider.write(uref, value)
 }
 
 pub fn get_bids<P>(provider: &mut P) -> Result<Bids, Error>
@@ -153,7 +152,11 @@ pub fn set_seigniorage_recipients_snapshot<P>(
 where
     P: StorageProvider + RuntimeProvider + ?Sized,
 {
-    write_to(provider, SEIGNIORAGE_RECIPIENTS_SNAPSHOT_KEY, snapshot)
+    let key = provider
+        .named_keys_get(SEIGNIORAGE_RECIPIENTS_SNAPSHOT_KEY)
+        .ok_or(Error::MissingKey)?;
+    let uref = key.into_uref().ok_or(Error::InvalidKeyVariant)?;
+    provider.write_seigniorage_recipients_snapshot(uref, snapshot)
 }
 
 pub fn get_validator_slots<P>(provider: &mut P) -> Result<usize, Error>